Clarify the mission on updates, state, and scope
Updates are the quintessential operation. Add further clarification
on expectations related to state, and factory reset.

In particular, it should be easier for people to understand what is
and isn't candidates for including in a bootable container definition,
images, artifacts and implementing in the underlying components.

Just like containers have scope so do bootable containers, and certain
deployment choices don't belong in this technology.

Signed-off-by: Stef Walter <stefw@redhat.com>